refactor(api): move unpublish route from lib to src
