+++ api/lib/application/sessions/index.js
@@ -69,34 +69,6 @@ const register = async function (server) {
         tags: ['api', 'session', 'publication'],
       },
     },
-    {
-      method: 'PATCH',
-      path: '/api/admin/sessions/{id}/unpublish',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.sessionId,
-          }),
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: sessionController.unpublish,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Dépublie toutes les certifications courses d'une session",
-        ],
-        tags: ['api', 'session', 'publication'],
-      },
-    },
   ]);
 };
