+++ api/tests/certification/session-management/unit/application/session-publication-route_test.js
@@ -59,5 +59,47 @@ describe('Unit | Application | Sessions | Routes', function () {
         expect(response.statusCode).to.equal(403);
       });
     });
+
+    describe('PATCH /api/admin/sessions/{sessionId}/unpublish', function () {
+      it('should exist', async function () {
+        // given
+        sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+        sinon.stub(sessionPublicationController, 'unpublish').returns('ok');
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const response = await httpTestServer.request('PATCH', '/api/admin/sessions/1/unpublish');
+
+        // then
+        expect(response.statusCode).to.equal(200);
+      });
+
+      it('return forbidden access if user has METIER role', async function () {
+        // given
+        sinon
+          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+          .withArgs([
+            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleSupport,
+          ])
+          .callsFake(
+            () => (request, h) =>
+              h
+                .response({ errors: new Error('forbidden') })
+                .code(403)
+                .takeover(),
+          );
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const response = await httpTestServer.request('PATCH', '/api/admin/sessions/1/unpublish');
+
+        // then
+        expect(response.statusCode).to.equal(403);
+      });
+    });
   });
 });
